+++ src/main/java/svenhjol/charmony/core/base/Environment.java
@@ -20,12 +20,12 @@ public final class Environment {
     }
 
     public static boolean isDebugMode() {
-        return Charmony.instance().tryFeature(Core.class)
+        return Charmony.instance().trySidedFeature(Core.class)
             .map(Core::debugMode).orElse(false);
     }
 
     public static boolean isMixinDisableMode() {
-        return Charmony.instance().tryFeature(Core.class)
+        return Charmony.instance().trySidedFeature(Core.class)
             .map(Core::mixinsDisabled).orElse(false);
     }
 
@@ -35,7 +35,7 @@ public final class Environment {
      */
     @net.fabricmc.api.Environment(EnvType.CLIENT)
     public static boolean usesCharmonyServer() {
-        return Charmony.instance().tryFeature(svenhjol.charmony.core.client.features.core.Core.class)
+        return Charmony.instance().trySidedFeature(svenhjol.charmony.core.client.features.core.Core.class)
             .map(c -> c.handlers.usesCharmonyServer()).orElse(false);
     }
 }